+++ src/Agent/NewRelic/Agent/Core/Aggregators/MetricAggregator.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using NewRelic.Agent.Core.DataTransport;
 using NewRelic.Agent.Core.Events;
-using NewRelic.Agent.Core.Metric;
 using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.SharedInterfaces;
 using NewRelic.Agent.Core.Time;
