+++ src/Agent/NewRelic/Agent/Core/Api/AgentApi.cs
@@ -4,7 +4,7 @@
 #nullable enable
 
 using NewRelic.Agent.Api;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Core.CodeAttributes;
 using NewRelic.Core.Logging;
 using System;
