+++ src/Agent/NewRelic/Agent/Core/Api/TransactionBridgeApi.cs
@@ -3,7 +3,7 @@
 
 using System;
 using NewRelic.Agent.Api;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Configuration;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using NewRelic.Core.Logging;
