+++ src/Agent/NewRelic/Agent/Core/DataTransport/ConnectionHandler.cs
@@ -6,7 +6,7 @@ using NewRelic.Agent.Core.AgentHealth;
 using NewRelic.Agent.Core.Configuration;
 using NewRelic.Agent.Core.Events;
 using NewRelic.Agent.Core.Labels;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.Utilities;
 using NewRelic.Agent.Core.Utilization;
 using NewRelic.Core.Logging;
