+++ src/Agent/NewRelic/Agent/Core/Metrics/RegexRuleExtensions.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using NewRelic.Agent.Configuration;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Helpers;
 
 namespace NewRelic.Agent.Core.Metrics
