+++ src/Agent/NewRelic/Agent/Core/Segments/MethodSegmentData.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using static NewRelic.Agent.Core.WireModels.MetricWireModel;
 using NewRelic.Agent.Core.Aggregators;
 using NewRelic.Agent.Core.Time;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Configuration;
 
 namespace NewRelic.Agent.Core.Segments
