+++ src/Agent/NewRelic/Agent/Core/Transactions/TransactionName.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using System;
 using System.Text;
