+++ tests/Agent/UnitTests/CompositeTests/ApiSupportabilityMetricTests.cs
@@ -5,7 +5,7 @@ using NewRelic.Agent.Api;
 using NewRelic.Agent.Core;
 using NewRelic.Agent.Configuration;
 using NewRelic.Agent.Core.Api;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using NUnit.Framework;
 using System;
