+++ tests/Agent/UnitTests/Core.UnitTest/Api/TransactionBridgeApiTests.cs
@@ -5,7 +5,7 @@
 using System;
 using NewRelic.Agent.Api;
 using NewRelic.Agent.Configuration;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NUnit.Framework;
 using Telerik.JustMock;
