+++ tests/Agent/UnitTests/Core.UnitTest/Metrics/CATSupportabilityMetricCounterTests.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.WireModels;
 using NUnit.Framework;
 using System.Collections.Generic;
