+++ tests/Agent/UnitTests/Core.UnitTest/Metrics/MetricNamesTests.cs
@@ -3,7 +3,7 @@
 
 using Grpc.Core;
 using NewRelic.Agent.Core.AgentHealth;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.Samplers;
 using NewRelic.Agent.Core.Transformers.TransactionTransformer;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
