+++ tests/Agent/UnitTests/Core.UnitTest/Transformers/ThreadStatsSampleTransformerTests.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 using NewRelic.Agent.Core.Aggregators;
-using NewRelic.Agent.Core.Metric;
+using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.Metrics;
 using NewRelic.Agent.Core.Samplers;
 using NewRelic.Agent.Core.WireModels;
